Rename module `Internal.Rounding`.
Rename module `Internal.Rounding` to `Internal.Numeric.Rounding`.
